Transition to react vite. Add re-rendering logic for links
