+++ src/index.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import App from './App.jsx';
+import reportWebVitals from './reportWebVitals.js';
+
+/**
+ * React Boostrap Config
+ */
+import "react-bootstrap/dist/react-bootstrap.js"
+import "../node_modules/bootstrap/dist/css/bootstrap.css"
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
+
+reportWebVitals();
